+++ packages/flame/test/_resources/load_image.dart
