+++ packages/flame/test/sprite_test.dart
@@ -6,7 +6,7 @@ import 'package:flame_test/flame_test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '_resources/images.dart';
+import '_resources/load_image.dart';
 
 void main() {
   group('Sprite', () {
